[fix]: StoppableLock 에서 ChannelLock으로 메커니즘을 변경
어떤 동적인 채널의 아이템이 들어있고, 이 채널에 아이템이 들어가는 동시성을 보장하기 위해서는 락이 제거될 때 위험하다. 동적인 락을 들고 있는 컬렉션에서 제거되면 비동기적인 처리에 의해 아이템이 미쳐 제거되기전에 채널 락이 사라지는 경우, 이 아이템은 영영 삭제가 불가능하기 때문이다.

이를 위해서 입장/퇴장할 때 락의 운용 방식을 바꾸어, 채널 삭제 요청에 의해 채널이 삭제되더라도 채널에 남아있는 아이템들이 모두 제거된 다음 이 락을 들고 있는 컬렉션에게 제거 요청 메서드를 날려 안전하게 모든 아이템이 제거되는 것을 보장한다.

softeer5th#327
