+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/bus/simulator/BusSubscriptionLockManager.java
@@ -2,7 +2,7 @@ package com.ddbb.dingdong.infrastructure.bus.simulator;
 
 
 import com.ddbb.dingdong.domain.transportation.repository.BusScheduleQueryRepository;
-import com.ddbb.dingdong.infrastructure.lock.StoppableLock;
+import com.ddbb.dingdong.infrastructure.lock.ChannelLock;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,28 +16,27 @@ import java.util.concurrent.ConcurrentHashMap;
 @RequiredArgsConstructor
 public class BusSubscriptionLockManager {
     private final BusScheduleQueryRepository busScheduleQueryRepository;
-    private final Map<Long, StoppableLock> locks = new ConcurrentHashMap<>();
+    private final Map<Long, ChannelLock> locks = new ConcurrentHashMap<>();
 
     @PostConstruct
     private void init() {
         List<Long> busSchedules = busScheduleQueryRepository.findLiveBusSchedule();
         for (Long busScheduleId : busSchedules) {
-            locks.put(busScheduleId, new StoppableLock());
+            locks.put(busScheduleId, new ChannelLock());
         }
     }
 
-
-    public Optional<StoppableLock> getLock(long busScheduleId) {
-        StoppableLock stampedLock = locks.get(busScheduleId);
+    public Optional<ChannelLock> getLock(long busScheduleId) {
+        ChannelLock stampedLock = locks.get(busScheduleId);
         return Optional.ofNullable(stampedLock);
     }
 
     public void addLock(long busScheduleId) {
-        locks.computeIfAbsent(busScheduleId, id -> new StoppableLock());
+        locks.computeIfAbsent(busScheduleId, id -> new ChannelLock());
     }
 
     public void removeLock(long busScheduleId) {
-        StoppableLock lock = locks.get(busScheduleId);
+        ChannelLock lock = locks.get(busScheduleId);
         if (lock != null) {
             lock.stopAndWait();
             locks.remove(busScheduleId);
